Move tests to storage
